requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djongo==1.3.6
dnspython==2.3.0
drf-yasg==1.21.5
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
mongoengine==0.27.0
openapi-codec==1.3.2
packaging==23.1
pika==1.2.0
PyJWT==2.6.0
pymongo==3.12.3
python-dotenv==0.19.2
pytz==2023.3
requests==2.29.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
simplejson==3.19.1
sqlparse==0.2.4
uritemplate==4.1.1
urllib3==1.26.15
